+++ NEWS.md
@@ -1,3 +1,9 @@
+# representr 0.1.2
+
+- Remove dependencies on `blink` and `stringdist` packages which are no longer supported on CRAN
+- Don't test long running examples
+
+
 # representr 0.1.1
 
 - Updated to include support for noLD (without long double)